Cleanup timing across the engine
- Moved time functions from WindowBase -> TimeKeeper
- WindowBase now has a `time_keeper` Property
- The Window's TimeKeeper instance is now passed to RigidBodySimulation
- Implemented interpolation in the rigid body controller
- Switched `double` to `float` for all delta time and step values (unnecessary precision)
- Removed `WindowBase::total_time()`, `WindowBase::delta_time()` etc. replaced with `time_keeper->total_elapsed_seconds()` etc.

+++ disabled/path_following.cpp
@@ -45,7 +45,7 @@ public:
         return true;
     }
 
-    void update(double dt) {
+    void update(float dt) {
         set_velocity(follower_->steer_to_path());
         actor()->move_to(position() + (velocity() * dt));
     }
@@ -87,7 +87,7 @@ private:
         return true;
     }
 
-    void do_step(double dt) {
+    void do_step(float dt) {
         if(initialized()) {
             car_->update(dt);
         }
